+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "start": "node server.js"
+    "start": "node src/index.js"
   },
   "keywords": [],
   "author": "",
