+++ src/requestHandler.js
@@ -1,65 +1,3 @@
-import http from 'http';
-import url from 'url';
-import fs from 'fs';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import { dirname } from 'path';
-
-// Helper functions to get the current file's directory
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-// Define the directory to store data files
-const DATA_DIR = path.join(__dirname, 'data');
-
-// Ensure the data directory exists
-if (!fs.existsSync(DATA_DIR)) {
-    fs.mkdirSync(DATA_DIR);
-}
-
-class FileManager {
-    constructor(dataDir) {
-        this.dataDir = dataDir;
-    }
-
-    getTableFilePath(tableName) {
-        return path.join(this.dataDir, `${tableName}.json`);
-    }
-
-    readFile(filePath) {
-        return new Promise((resolve, reject) => {
-            fs.readFile(filePath, (err, data) => {
-                if (err) {
-                    return reject(err);
-                }
-                resolve(JSON.parse(data));
-            });
-        });
-    }
-
-    writeFile(filePath, data) {
-        return new Promise((resolve, reject) => {
-            fs.writeFile(filePath, JSON.stringify(data), (err) => {
-                if (err) {
-                    return reject(err);
-                }
-                resolve();
-            });
-        });
-    }
-
-    deleteFile(filePath) {
-        return new Promise((resolve, reject) => {
-            fs.unlink(filePath, (err) => {
-                if (err) {
-                    return reject(err);
-                }
-                resolve();
-            });
-        });
-    }
-}
-
 class RequestHandler {
     constructor(fileManager) {
         this.fileManager = fileManager;
@@ -214,48 +152,4 @@ class RequestHandler {
     }
 }
 
-class Server {
-    constructor(port, fileManager, requestHandler) {
-        this.port = port || 3000;
-        this.fileManager = fileManager;
-        this.requestHandler = requestHandler;
-        this.server = http.createServer((req, res) => this.handleRequest(req, res));
-    }
-
-    handleRequest(req, res) {
-        const parsedUrl = url.parse(req.url, true);
-        const { pathname } = parsedUrl;
-        const pathParts = pathname.split('/').filter(Boolean);
-
-        if (req.method === 'POST' && pathParts[0] === 'data' && pathParts[1] === 'store') {
-            this.requestHandler.handleStoreData(req, res);
-        } else if (req.method === 'GET' && pathParts.length === 3 && pathParts[0] === 'data' && pathParts[2] === 'all') {
-            this.requestHandler.handleGetAllData(req, res, pathParts[1]);
-        } else if (req.method === 'GET' && pathParts.length === 4 && pathParts[0] === 'data' && pathParts[2] === 'entry') {
-            this.requestHandler.handleGetDataByEntryId(req, res, pathParts[1], pathParts[3]);
-        } else if (req.method === 'PUT' && pathParts.length === 4 && pathParts[0] === 'data' && pathParts[2] === 'update') {
-            this.requestHandler.handleUpdateData(req, res, pathParts[1], pathParts[3]);
-        } else if (req.method === 'DELETE' && pathParts.length === 4 && pathParts[0] === 'data' && pathParts[2] === 'delete') {
-            this.requestHandler.handleDeleteDataByEntryId(req, res, pathParts[1], pathParts[3]);
-        } else if (req.method === 'DELETE' && pathParts.length === 3 && pathParts[0] === 'data' && pathParts[2] === 'delete') {
-            this.requestHandler.handleDeleteTable(req, res, pathParts[1]);
-        } else {
-            res.writeHead(404, { 'Content-Type': 'application/json' });
-            res.end(JSON.stringify({ error: 'Endpoint not found' }));
-        }
-    }
-
-    start() {
-        this.server.listen(this.port, () => {
-            console.log(`Server is running on port ${this.port}`);
-        });
-    }
-}
-
-const fileManager = new FileManager(DATA_DIR);
-const requestHandler = new RequestHandler(fileManager);
-const server = new Server(process.env.PORT, fileManager, requestHandler);
-
-server.start();
-
-export default server;
+export default RequestHandler;
